+++ example/a.ts
@@ -1,46 +1,80 @@
 import { t } from '../src/type'
 import { Elysia } from '../src/elysia'
 import { compileHandler } from '../src/compile'
-import { Validator } from '../src/schema/validator'
 
-const handler = compileHandler(
-	[
-		'/',
-		'POST',
-		({ body }) => {
-			return body.name
-		},
-		{
-			body: t.Array(
-				t.Object({
-					name: t.String(),
-					age: t.Number({
-						default: 1
-					})
-				})
-			)
-		},
-		new Elysia()
-	],
-	new Elysia()
-)
+function gc() {
+	if (typeof Bun !== 'undefined') Bun.gc(true)
+	else if (typeof global.gc === 'function') global.gc()
+}
 
-await handler({
-	set: {
-		status: 200,
-		headers: {}
-	},
-	request: new Request('http://localhost?a=b', {
-		method: 'GET',
-		headers: {
-			'content-type': 'application/json'
-		},
-		body: JSON.stringify([
-			{
-				name: 'q'
-			}
-		])
-	})
-})
-	.then((res) => res.text())
-	.then(console.log)
+function memoryUsage() {
+	if (typeof Bun !== 'undefined') {
+		const { memoryUsage } = require('bun:jsc')
+
+		gc()
+		return memoryUsage().current
+	}
+
+	gc()
+	return process.memoryUsage().heapUsed
+}
+
+const app = new Elysia()
+
+const total = 100_000
+const stacks = <any[]>Array(total)
+const m1 = memoryUsage()
+const t1 = performance.now()
+
+for (let i = 0; i <= total; i++)
+	stacks.push(
+		compileHandler(
+			[
+				'/',
+				'POST',
+				({ body }) => 'q',
+				{
+					body: t.Array(
+						t.Object({
+							name: t.String(),
+							age: t.Number({
+								default: 1
+							})
+						})
+					)
+				},
+				app
+			],
+			app
+		)
+	)
+
+const t2 = performance.now()
+gc()
+const m2 = memoryUsage()
+
+// console.log('\n\n\n')
+console.log('Elysia 2 schema with 45 types x100,000')
+console.log('Time:', (t2 - t1).toFixed(2), 'ms')
+console.log('Memory usage:', ((m2 - m1) / 1024 / 1024).toFixed(2), 'MB')
+// console.log('\n\n\n\n\n\n')
+
+// await handler({
+// 	set: {
+// 		status: 200,
+// 		headers: {}
+// 	},
+// 	request: new Request('http://localhost?a=b', {
+// 		method: 'GET',
+// 		headers: {
+// 			'content-type': 'application/json'
+// 		},
+// 		body: JSON.stringify([
+// 			{
+// 				name: 'q'
+// 			}
+// 		])
+// 	})
+// })
+// 	.then((res) => res.text())
+// 	.then(console.log)